chore(router): remove legacy `createRouter` implementation
