+++ src/router.ts
@@ -1,9 +1,122 @@
-import type { HTTPMethod, RequestContext, RouteEndpoint, RoutePattern, RouterConfig, GetHttpHandlers } from "./types.js"
-import { createRoutePattern } from "./endpoint.js"
-import { getBody, getHeaders, getRouteParams, getSearchParams } from "./context.js"
-import { executeGlobalMiddlewares, executeMiddlewares } from "./middlewares.js"
-import { RouterError, statusText } from "./error.js"
 import { isRouterError, isSupportedMethod } from "./assert.js"
+import { getBody, getHeaders, getRouteParams, getSearchParams } from "./context.js"
+import { RouterError, statusText } from "./error.js"
+import { executeGlobalMiddlewares, executeMiddlewares } from "./middlewares.js"
+import type { GetHttpHandlers, HTTPMethod, RouteEndpoint, RoutePattern, RouterConfig } from "./types.js"
+
+interface TrieNode {
+    statics: Map<string, TrieNode>
+    param?: { name: string; node: TrieNode }
+    endpoints: Map<HTTPMethod, RouteEndpoint>
+}
+
+const createNode = (): TrieNode => ({
+    statics: new Map(),
+    endpoints: new Map(),
+})
+
+const insert = (root: TrieNode, endpoint: RouteEndpoint) => {
+    if (!root || !endpoint) return
+    let node = root
+    const segments = endpoint.route === "/" ? [] : endpoint.route.split("/").filter(Boolean)
+    for (const segment of segments) {
+        if (segment.startsWith(":")) {
+            const name = segment.slice(1)
+            if (!node.param) {
+                node.param = { name, node: createNode() }
+            } else if (node.param.name !== name) {
+                throw new RouterError(
+                    "BAD_REQUEST",
+                    `Conflicting in the route by the dynamic segment "${node.param.name}" and "${name}"`
+                )
+            }
+            node = node.param.node
+        } else {
+            if (!node.statics.has(segment)) {
+                node.statics.set(segment, createNode())
+            }
+            node = node.statics.get(segment)!
+        }
+    }
+    if (node.endpoints.has(endpoint.method)) {
+        throw new RouterError("BAD_REQUEST", `Duplicate endpoint for ${endpoint?.method} ${endpoint?.route}`)
+    }
+    node.endpoints.set(endpoint.method, endpoint)
+}
+
+const search = (root: TrieNode, pathname: string, method: HTTPMethod) => {
+    let node = root
+    const params = {} as Record<string, string>
+    const segments = pathname === "/" ? [] : pathname.split("/").filter(Boolean)
+    for (const segment of segments) {
+        if (node?.statics.has(segment)) {
+            node = node.statics.get(segment)!
+        } else if (node?.param) {
+            params[node.param.name] = decodeURIComponent(segment)
+            node = node.param.node
+        } else {
+            throw new RouterError("NOT_FOUND", `No route found for path: ${pathname}`)
+        }
+    }
+    if (!node.endpoints.has(method)) {
+        throw new RouterError("NOT_FOUND", `No route found for path: ${pathname}`)
+    }
+    return { endpoint: node.endpoints.get(method)!, params }
+}
+
+const handleError = async (error: unknown, request: Request, config: RouterConfig) => {
+    if (config.onError) {
+        try {
+            const response = await config.onError(error as Error | RouterError, request)
+            return response
+        } catch {
+            return Response.json(
+                { message: "A critical failure occurred during error handling" },
+                { status: 500, statusText: statusText.INTERNAL_SERVER_ERROR }
+            )
+        }
+    }
+    if (isRouterError(error)) {
+        const { message, status, statusText } = error
+        return Response.json({ message }, { status, statusText })
+    }
+    return Response.json({ message: "Internal Server Error" }, { status: 500, statusText: statusText.INTERNAL_SERVER_ERROR })
+}
+
+const handleRequest = async (method: HTTPMethod, request: Request, config: RouterConfig, root: TrieNode) => {
+    try {
+        if (!isSupportedMethod(request.method)) {
+            throw new RouterError("METHOD_NOT_ALLOWED", `The HTTP method '${request.method}' is not supported`)
+        }
+        const globalRequest = await executeGlobalMiddlewares(request, config.middlewares)
+        if (globalRequest instanceof Response) return globalRequest
+
+        const url = new URL(globalRequest.url)
+        const pathnameWithBase = url.pathname
+        if (globalRequest.method !== method) {
+            throw new RouterError("METHOD_NOT_ALLOWED", `The HTTP method '${globalRequest.method}' is not allowed`)
+        }
+        const { endpoint, params } = search(root, pathnameWithBase, method)
+        if (endpoint.method !== globalRequest.method) {
+            throw new RouterError("METHOD_NOT_ALLOWED", `The HTTP method '${globalRequest.method}' is not allowed`)
+        }
+        const dynamicParams = getRouteParams(params, endpoint.config)
+        const body = await getBody(globalRequest, endpoint.config)
+        const searchParams = getSearchParams(globalRequest.url, endpoint.config)
+        const headers = getHeaders(globalRequest)
+        const context = {
+            params: dynamicParams,
+            searchParams,
+            headers,
+            body,
+        }
+        await executeMiddlewares(globalRequest, context, endpoint.config.middlewares)
+        const response = await endpoint.handler(globalRequest, context)
+        return response
+    } catch (error) {
+        return handleError(error, request, config)
+    }
+}
 
 /**
  * Creates the entry point for the server, handling the endpoints defined in the router.
@@ -18,82 +131,16 @@ export const createRouter = <const Endpoints extends RouteEndpoint[]>(
     endpoints: Endpoints,
     config: RouterConfig = {}
 ): GetHttpHandlers<Endpoints> => {
+    const root = createNode()
     const server = {} as GetHttpHandlers<Endpoints>
-    const groups = new Map<HTTPMethod, RouteEndpoint[]>()
+    const methods = new Set<HTTPMethod>()
     for (const endpoint of endpoints) {
-        if (!groups.has(endpoint.method)) {
-            groups.set(endpoint.method, [])
-        }
-        groups.get(endpoint.method)?.push(endpoint)
+        const withBasePath = config.basePath ? `${config.basePath}${endpoint.route}` : endpoint.route
+        insert(root, { ...endpoint, route: withBasePath as RoutePattern })
+        methods.add(endpoint.method)
     }
-    for (const method of groups.keys()) {
-        server[method as keyof typeof server] = async (request: Request) => {
-            return await matchRoute(request, method, groups, config)
-        }
+    for (const method of methods) {
+        server[method as keyof typeof server] = (request: Request) => handleRequest(method, request, config, root)
     }
     return server
-}
-
-const matchRoute = async (
-    request: Request,
-    method: HTTPMethod,
-    groups: Map<HTTPMethod, RouteEndpoint[]>,
-    config: RouterConfig
-) => {
-    try {
-        if (!isSupportedMethod(request.method)) {
-            throw new RouterError("METHOD_NOT_ALLOWED", `The HTTP method '${request.method}' is not supported`)
-        }
-        if (!groups.has(request.method)) {
-            throw new RouterError("METHOD_NOT_ALLOWED", `The HTTP method '${request.method}' is not allowed`)
-        }
-        const globalRequest = await executeGlobalMiddlewares(request, config.middlewares)
-        if (globalRequest instanceof Response) {
-            return globalRequest
-        }
-        const url = new URL(globalRequest.url)
-        const pathname = url.pathname
-        const endpoint = groups.get(method)?.find((endpoint) => {
-            const withBasePath = config.basePath ? `${config.basePath}${endpoint.route}` : endpoint.route
-            const regex = createRoutePattern(withBasePath as RoutePattern)
-            return regex.test(pathname)
-        })
-        if (endpoint) {
-            if (globalRequest.method !== endpoint.method) {
-                throw new RouterError("METHOD_NOT_ALLOWED", `The HTTP method '${globalRequest.method}' is not allowed`)
-            }
-            const withBasePath = config.basePath ? `${config.basePath}${endpoint.route}` : endpoint.route
-            const body = await getBody(globalRequest, endpoint.config)
-            const params = getRouteParams(withBasePath as RoutePattern, pathname, endpoint.config)
-            const searchParams = getSearchParams(globalRequest.url, endpoint.config)
-            const headers = getHeaders(globalRequest)
-            const context = {
-                params,
-                searchParams,
-                headers,
-                body,
-            } as RequestContext<Record<string, string>, typeof endpoint.config>
-            await executeMiddlewares(globalRequest, context, endpoint.config.middlewares)
-            const response = await endpoint.handler(globalRequest, context)
-            return response
-        }
-        throw new RouterError("NOT_FOUND", "Not Found")
-    } catch (error) {
-        if (config.onError) {
-            try {
-                const response = await config.onError(error as Error | RouterError, request)
-                return response
-            } catch {
-                return Response.json(
-                    { message: "A critical failure occurred during error handling" },
-                    { status: 500, statusText: statusText.INTERNAL_SERVER_ERROR }
-                )
-            }
-        }
-        if (isRouterError(error)) {
-            const { message, status, statusText } = error
-            return Response.json({ message }, { status, statusText })
-        }
-        return Response.json({ message: "Internal Server Error" }, { status: 500, statusText: statusText.INTERNAL_SERVER_ERROR })
-    }
 }
